src/main/java/ru/job4j/io/Config.java
package ru.job4j.io;

import java.io.BufferedReader;
import java.io.FileReader;
import java.io.IOException;
import java.util.HashMap;
import java.util.List;
import java.util.Map;
import java.util.StringJoiner;

public class Config {

    private final String path;
    private final Map<String, String> values = new HashMap<String, String>();

    public Config(final String path) {
        this.path = path;
    }

    public void load() {
        try (BufferedReader read = new BufferedReader(new FileReader(this.path))) {
            List<String> outList = read.lines().filter(line -> !line.split("#")[0].isBlank()).toList();
            for (String line: outList) {
                String lineWithoutComment = line.split("#")[0];
                int equalIndex = lineWithoutComment.indexOf("=");
                if (equalIndex <= 0 || equalIndex == lineWithoutComment.length() - 1) {
                    throw new IllegalArgumentException();
                }
                String key = lineWithoutComment.substring(0, equalIndex);
                String value = lineWithoutComment.substring(equalIndex + 1);
                values.put(key, value);
            }
        } catch (IOException e) {
            e.printStackTrace();
        }
    }

    public String value(String key) {
        return values.get(key);

    }

    @Override
    public String toString() {
        StringJoiner out = new StringJoiner(System.lineSeparator());
        try (BufferedReader read = new BufferedReader(new FileReader(this.path))) {
            read.lines().forEach(out::add);
        } catch (IOException e) {
            e.printStackTrace();
        }
        return out.toString();
    }

    public static void main(String[] args) {
        System.out.println(new Config("app.properties"));
    }

}
